+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Refactor URL key handling
 * Remove deprecated classes and methods
+* Add techdivision/import#162
 * Add techdivision/import-product#146
 * Add techdivision/import-cli-simple#216
 * Add techdivision/import-configuration-jms#25